+++ src/main/java/com/autobots/automanager/controller/TelephoneController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.autobots.automanager.entity.Telephone;
+import com.autobots.automanager.models.ErrorTreatment;
+import com.autobots.automanager.models.TelephoneLinkAdder;
 import com.autobots.automanager.services.TelephoneService;
 
 
@@ -28,33 +30,109 @@ public class TelephoneController {
 	@Autowired
 	private TelephoneService service;
 	
-	@GetMapping("/telephones")
+	@Autowired
+	private TelephoneLinkAdder linkAdder;
+	
+	@GetMapping("/clients")
 	public ResponseEntity<List<Telephone>> getAllTelephones(){
 		List<Telephone> allTelephones = service.findAll();
-		return new ResponseEntity<>(allTelephones, HttpStatus.ACCEPTED);
+		if(allTelephones.isEmpty()) {
+			ResponseEntity<List<Telephone>> response = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			return response;
+		} else {
+			linkAdder.addLink(allTelephones);
+			ResponseEntity<List<Telephone>> response = new ResponseEntity<>(allTelephones, HttpStatus.ACCEPTED);
+			return response;
+		}
 	}
 	
-	@GetMapping("/telephone/{id}")
+	@GetMapping("/client/{id}")
 	public ResponseEntity<Telephone> getTelephone(@PathVariable Long id){
+		List<Telephone> allTelephones = service.findAll();
 		Telephone telephone = service.findById(id);
-		return new ResponseEntity<>(telephone, HttpStatus.ACCEPTED);
+		if(telephone == null) {
+			ResponseEntity<Telephone> response = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			return response;
+		} else {
+			linkAdder.addLink(telephone);
+			ResponseEntity<Telephone> response = new ResponseEntity<>(telephone, HttpStatus.FOUND);
+			return response;
+		}
+
 	}
 	
 	@PostMapping("/register")
 	public ResponseEntity<?> insertNewTelephone(@RequestBody Telephone obj){
-		service.insert(obj);
-		return new ResponseEntity<>("successful request", HttpStatus.ACCEPTED);
+        ErrorTreatment<Telephone> errorTreatment = new ErrorTreatment<>();
+        
+		HttpStatus status;
+		String responseString;
+        
+        errorTreatment.checkCopyItemToList(service.findAll(), obj);
+        
+        if(errorTreatment.getHasCopyError()) {
+        	
+        	responseString = "Object has a copy. Object:" + errorTreatment.getHasCopyError();
+            status = HttpStatus.CONFLICT;
+            
+        }else if(errorTreatment.getIsNullError()){
+        	
+            status = HttpStatus.NOT_FOUND;
+            responseString = "Body cannot be null";    
+            
+        }else {
+        	
+            status = HttpStatus.ACCEPTED;
+            responseString = "Successful request";
+            
+        	service.insert(obj);
+        }
+        return new ResponseEntity<>(responseString, status);
 	}
 
 	@PutMapping("/update")
 	public ResponseEntity<?> updateTelephone(@RequestBody Telephone obj){
-		service.update(obj);
-		return new ResponseEntity<>("successful request", HttpStatus.ACCEPTED);	
+		HttpStatus status;
+		String responseString;
+		
+		ErrorTreatment<Telephone> errorTreatment = new ErrorTreatment<>();
+		errorTreatment.isObjectNull(obj);
+		
+		if(errorTreatment.getIsNullError()) {
+			
+			responseString = errorTreatment.getErrorLog();
+			status = HttpStatus.NOT_FOUND;
+			
+		}else {
+			
+			responseString = "Successful request";
+			status = HttpStatus.ACCEPTED;
+			
+			service.update(obj);
+		}
+		
+		return new ResponseEntity<>(responseString, status);
 	}
 	
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<?> deleteTelephone(@PathVariable Long id){
-		service.delete(id);
-		return new ResponseEntity<>("successful request", HttpStatus.ACCEPTED);	
+		HttpStatus status;
+		String responseString;
+		
+		ErrorTreatment<Telephone> errorTreatment = new ErrorTreatment<>();
+		
+		Telephone telephone = service.findById(id);
+		
+		if(errorTreatment.isObjectNull(telephone)) {
+			status = HttpStatus.NOT_FOUND;
+			responseString = "Object not found";
+		}else {
+			status = HttpStatus.ACCEPTED;
+			responseString = "Successful request";
+			service.delete(id);
+			
+		}
+		
+		return new ResponseEntity<>(responseString, status);
 	}
 }
